Move the apitest package into oteltest
